test(acc): split e2e recording tests into parallel per-file structure
Each recording gets its own file under test/e2e/acc/ so Bun runs them
in separate worker threads in parallel. Adds TestLogger helper to
buffer and flush output as a single grouped block after each test
completes, preventing interleaved logs from concurrent workers.
